Spliting the screens into seperate files
